housekeeping: reorganize transformRenderWithShip into VSClientGameUtils
